@@ -1,33 +1,22 @@
 {
     "name": "league/flysystem-webdav",
-    "description": "Flysystem adapter for WebDAV",
+    "description": "WebDAV filesystem adapter for Flysystem.",
+    "keywords": ["flysystem", "filesystem", "webdav", "files", "file"],
+    "autoload": {
+        "psr-4": {
+                "League\\Flysystem\\WebDAV\\": ""
+        }
+    },
+    "require": {
+        "php": "^8.0.2",
+        "league/flysystem": "^3.6.0",
+        "sabre/dav": "^4.6.0"
+    },
     "license": "MIT",
     "authors": [
         {
             "name": "Frank de Jonge",
-            "email": "info@frenky.net"
+            "email": "info@frankdejonge.nl"
         }
-    ],
-    "require": {
-        "php": ">=5.4.0",
-        "league/flysystem": "~1.0",
-        "sabre/dav": "~2.0"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "~4.0",
-        "mockery/mockery": "~0.9"
-    },
-    "autoload": {
-        "psr-4": {
-            "League\\Flysystem\\WebDAV\\": "src"
-        }
-    },
-    "config": {
-        "bin-dir": "bin"
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.0-dev"
-        }
-    }
+    ]
 }
