@@ -3,7 +3,7 @@
     "type": "library",
     "description": "WebDAV Framework for PHP",
     "keywords": ["Framework", "WebDAV", "CalDAV", "CardDAV", "iCalendar"],
-    "homepage": "http://code.google.com/p/sabredav/",
+    "homepage": "http://sabre.io/",
     "license" : "BSD-3-Clause",
     "authors": [
         {
@@ -14,10 +14,12 @@
         }
     ],
     "require": {
-        "php": ">=5.4.1",
-        "sabre/vobject" : "~3.3",
-        "sabre/event" : "~1.0.0",
-        "sabre/http" : "~2.0.2",
+        "php": "^7.1.0 || ^8.0",
+        "sabre/vobject": "^4.2.1",
+        "sabre/event" : "^5.0",
+        "sabre/xml"  : "^2.0.1",
+        "sabre/http" : "^5.0.5",
+        "sabre/uri" : "^2.0",
         "ext-dom": "*",
         "ext-pcre": "*",
         "ext-spl": "*",
@@ -26,22 +28,30 @@
         "ext-ctype" : "*",
         "ext-date" : "*",
         "ext-iconv" : "*",
-        "ext-libxml" : "*"
+        "lib-libxml" : ">=2.7.0",
+        "psr/log": "^1.0 || ^2.0 || ^3.0",
+        "ext-json": "*"
     },
     "require-dev" : {
-        "phpunit/phpunit" : "~4.2",
-        "evert/phpdoc-md" : "~0.0.7"
+        "friendsofphp/php-cs-fixer": "^2.19",
+        "monolog/monolog": "^1.27 || ^2.0",
+        "phpstan/phpstan": "^0.12 || ^1.0",
+        "phpstan/phpstan-phpunit": "^1.0",
+        "phpunit/phpunit": "^7.5 || ^8.5 || ^9.6"
     },
     "suggest" : {
         "ext-curl" : "*",
-        "ext-pdo" : "*"
+        "ext-pdo" : "*",
+        "ext-imap": "*"
     },
     "autoload": {
         "psr-4" : {
-            "Sabre\\DAV\\"     : "lib/DAV/",
-            "Sabre\\DAVACL\\"  : "lib/DAVACL/",
-            "Sabre\\CalDAV\\"  : "lib/CalDAV/",
-            "Sabre\\CardDAV\\" : "lib/CardDAV/"
+            "Sabre\\" : "lib/"
+        }
+    },
+    "autoload-dev" : {
+        "psr-4" : {
+            "Sabre\\" : "tests/Sabre/"
         }
     },
     "support" : {
@@ -52,7 +62,20 @@
         "bin/sabredav",
         "bin/naturalselection"
     ],
-    "config" : {
-        "bin-dir" : "./bin"
+    "scripts": {
+        "phpstan": [
+            "phpstan analyse lib tests"
+        ],
+        "cs-fixer": [
+            "php-cs-fixer fix"
+        ],
+        "phpunit": [
+            "phpunit --configuration tests/phpunit.xml"
+        ],
+        "test": [
+            "composer phpstan",
+            "composer cs-fixer",
+            "composer phpunit"
+        ]
     }
 }
